successfully merged mapping and generating function
However, one key problem to fix: when the mapping file is parsed, it's parsed as a csv with string values, NOT as python objects, which RDFLib terms are.
So, in the current test care, 'cco.Person' is being parsed as just a string and not the rdflib term.
To fix this, we need to parse the csv AS a python file... it will be fun trying to figure it out